mmrag-techpdf/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mmrag-techpdf"
version = "0.1.0"
description = "Multimodal RAG system for technical PDFs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",  # Settings management
    "typer>=0.9",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "pymupdf>=1.23",           # PDF processing + table finder
    "pillow>=10.0",            # Image processing
    "numpy>=1.24",
    "faiss-cpu>=1.7",          # Vector search (use faiss-gpu on CUDA machines)
    "sentence-transformers>=2.2",  # Embeddings + CrossEncoder reranker
    "transformers>=4.30",      # BLIP captioning + local LLM
    "torch>=2.0",              # PyTorch backend
    "rank-bm25>=0.2",          # BM25 sparse retrieval
    "openai>=1.0",             # OpenAI Chat Completions API
    "python-multipart>=0.0.6", # FastAPI file uploads
]

[project.scripts]
mmrag = "mmrag.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]